+++ packages/f-checkout/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.29.0
+------------------------------
+*December 23, 2020*
+
+### Added
+- `setup` test helper file.
+- Tests for `Checkout` component methods.
+
+
 v0.28.0
 ------------------------------
 *December 22, 2020*